# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Treasure KC Nexus API - LLC Discovery and Recruitment Platform
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface LlcFiling {
  id: number;
  name: string;
  /** @nullable */
  filingId?: string | null;
  state: string;
  status: string;
  filingDate: string;
  /** @nullable */
  agentName?: string | null;
  /** @nullable */
  agentAddress?: string | null;
  /** @nullable */
  city?: string | null;
  recruited: boolean;
  /** @nullable */
  recruitedAt?: string | null;
  /** @nullable */
  recruitNote?: string | null;
  createdAt: string;
}

export interface NewLlcsResponse {
  date: string;
  state: string;
  total: number;
  llcs: LlcFiling[];
}

export interface LlcListResponse {
  total: number;
  offset: number;
  limit: number;
  llcs: LlcFiling[];
}

export interface RecruitBody {
  /** @nullable */
  note?: string | null;
}

export interface ScrapeBody {
  /** State to scrape (KS, MO, or ALL) */
  state?: string;
  /**
   * Date to scrape for (YYYY-MM-DD), defaults to today
   * @nullable
   */
  date?: string | null;
}

export type SourceRefreshResultStatus =
  (typeof SourceRefreshResultStatus)[keyof typeof SourceRefreshResultStatus];

export const SourceRefreshResultStatus = {
  success: "success",
  failed: "failed",
} as const;

export interface SourceRefreshResult {
  state: string;
  sourceUrl: string;
  status: SourceRefreshResultStatus;
  found: number;
  stored: number;
  /** @nullable */
  error?: string | null;
  startedAt: string;
  completedAt: string;
}

export interface ScrapeResult {
  state: string;
  date: string;
  found: number;
  stored: number;
  message: string;
  sources: SourceRefreshResult[];
}

export type SourceHealthStatus =
  (typeof SourceHealthStatus)[keyof typeof SourceHealthStatus];

export const SourceHealthStatus = {
  success: "success",
  failed: "failed",
} as const;

export interface SourceHealth {
  state: string;
  sourceUrl: string;
  status: SourceHealthStatus;
  found: number;
  stored: number;
  /** @nullable */
  error?: string | null;
  lastRunAt: string;
}

export interface DashboardStats {
  totalLlcs: number;
  todayLlcs: number;
  kansasLlcs: number;
  missouriLlcs: number;
  recruitedCount: number;
  pendingRecruitment: number;
  /** @nullable */
  lastScrapeAt?: string | null;
  /** @nullable */
  lastRefreshAt?: string | null;
  sourceHealth?: SourceHealth[];
}

export interface CityCount {
  city: string;
  state: string;
  count: number;
}

export interface OutreachBody {
  /** Tone of message (friendly, professional, casual) */
  tone?: string;
}

export interface OutreachResponse {
  llcId: number;
  llcName: string;
  message: string;
  tone: string;
}

export type GetNewLlcsParams = {
  /**
   * State code (KS, MO, or ALL)
   */
  state?: string;
  /**
   * Filing date filter (YYYY-MM-DD), defaults to yesterday
   * @nullable
   */
  date?: string | null;
  /**
   * Maximum number of results to return
   */
  limit?: number;
  /**
   * Pagination offset
   */
  offset?: number;
};

export type GetFreshLlcsParams = {
  /**
   * State code (KS, MO, or ALL)
   */
  state?: string;
  /**
   * Filing date filter (YYYY-MM-DD), defaults to today
   * @nullable
   */
  date?: string | null;
  /**
   * Maximum number of results to return
   */
  limit?: number;
  /**
   * Pagination offset
   */
  offset?: number;
};

export type GetLlcsParams = {
  /**
   * Filter by state code (KS, MO)
   * @nullable
   */
  state?: string | null;
  /**
   * Search by LLC name or city
   * @nullable
   */
  search?: string | null;
  /**
   * Maximum number of results
   */
  limit?: number;
  /**
   * Pagination offset
   */
  offset?: number;
};

export type GetRecentActivityParams = {
  /**
   * Number of recent items
   */
  limit?: number;
};