# lib/api-zod/src/generated/types/dashboardStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Treasure KC Nexus API - LLC Discovery and Recruitment Platform
 * OpenAPI spec version: 0.1.0
 */
import type { SourceHealth } from "./sourceHealth";

export interface DashboardStats {
  totalLlcs: number;
  todayLlcs: number;
  kansasLlcs: number;
  missouriLlcs: number;
  recruitedCount: number;
  pendingRecruitment: number;
  /** @nullable */
  lastScrapeAt?: string | null;
  /** @nullable */
  lastRefreshAt?: string | null;
  sourceHealth?: SourceHealth[];
}