# lib/api-zod/src/generated/types/sourceHealth.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Treasure KC Nexus API - LLC Discovery and Recruitment Platform
 * OpenAPI spec version: 0.1.0
 */
import type { SourceHealthStatus } from "./sourceHealthStatus";

export interface SourceHealth {
  state: string;
  sourceUrl: string;
  status: SourceHealthStatus;
  found: number;
  stored: number;
  /** @nullable */
  error?: string | null;
  lastRunAt: string;
}