# lib/api-zod/src/generated/types/sourceRefreshResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Treasure KC Nexus API - LLC Discovery and Recruitment Platform
 * OpenAPI spec version: 0.1.0
 */
import type { SourceRefreshResultStatus } from "./sourceRefreshResultStatus";

export interface SourceRefreshResult {
  state: string;
  sourceUrl: string;
  status: SourceRefreshResultStatus;
  found: number;
  stored: number;
  /** @nullable */
  error?: string | null;
  startedAt: string;
  completedAt: string;
}